+++ forms-flow-api-utils/src/formsflow_api_utils/utils/__init__.py
@@ -1,6 +1,6 @@
 """This module holds general utility functions and helpers for the main package."""
 
-from .api_docs_models import authorization_list_model, submission_response
+from .api_docs_models import authorization_list_model, authorization_model, submission_response
 from .auth import auth, jwt
 from .constants import (
     ALLOW_ALL_APPLICATIONS,
@@ -21,6 +21,7 @@ from .constants import (
     NEW_APPLICATION_STATUS,
     REVIEWER_GROUP,
     HTTP_TIMEOUT,
+    authorization_doc_params,
 )
 from .enums import ApplicationSortingParameters
 from .permisions import (