+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Mark  items as `Added`, `Changed`, `Fixed`, `Modified`, `Removed`, `Untested Fea
 **formsflow-api**
 * Modified application get/update `/application/id` endpoint to support draft get and update.
 * Updated the anonymous draft POST API URL from `/draft/public/create` to `/public/draft`.
+* Updated the theme GET API URL from `/themes` to `/public/themes`.
 
 `Removed`